refactor(Follow): OutBoxStatus 패키지 변경 - event -> outbox
